+++ app/getStocksGraph.py
