+++ app/get_tickers.py
