+++ app/static/stocks.js
