+++ app/static/stocks.js.LICENSE.txt
