+++ app/templates/index.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html lang="en">
 	<head>
+		<meta charset="utf-8" />
 		<title>Chart the Stock Market</title>
 		<link rel="icon" href="https://link477.com/assets/images/link477.png"/>
 		<link href="myStyle.css" rel="stylesheet"/>
-		<meta charset="utf-8" />
 		<meta content="Tom Fishwick"/>
 		<meta name="viewport" content="width=device-width, initial-scale=1.0"/>
 		<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
@@ -32,7 +32,7 @@
 					</datalist>
 					<input autocomplete="on" list="stockcodes" id="stockCode" name="stockCode" class="form-control" required/>
 				</div>
-				<button class="btn btn-primary" id="btn">Submit</button>
+				<button class="btn btn-primary" id="btn" type="button">Submit</button>
 			</fieldset>
 		</div>
 		<script src="stocks.js" async></script>
