+++ webpack.config.js
@@ -4,6 +4,6 @@ module.exports = {
     entry: './src/stocks.js',
     output: {
         filename: 'stocks.js',
-        path: path.resolve(__dirname, 'static'),
+        path: path.resolve(__dirname, 'app', 'static'),
     },
 };
